♻️ refactor: 이메일 발송 인터페이스 위치 정리 및 알림 연동 일원화
