+++ app-api/src/test/java/com/tasteam/domain/group/service/GroupAuthServiceRateLimitTest.java
@@ -19,10 +19,10 @@ import com.tasteam.domain.member.repository.MemberRepository;
 import com.tasteam.global.config.DomainProperties;
 import com.tasteam.global.exception.business.BusinessException;
 import com.tasteam.global.exception.code.NotificationErrorCode;
-import com.tasteam.global.notification.email.EmailSender;
 import com.tasteam.global.ratelimit.RateLimitReason;
 import com.tasteam.global.ratelimit.RateLimitResult;
 import com.tasteam.global.ratelimit.RedisRateLimiter;
+import com.tasteam.infra.email.EmailSender;
 
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
